+++ src/commands/dev.ts
@@ -17,7 +17,7 @@ import { loadKit } from '../utils/kit'
 import { importModule } from '../utils/esm'
 import { overrideEnv } from '../utils/env'
 import type { NuxtDevContext, NuxtDevIPCMessage } from '../utils/dev'
-import { sharedArgs, legacyRootDirArgs } from './_shared'
+import { sharedArgs, envNameArgs, legacyRootDirArgs } from './_shared'
 
 const forkSupported = !isBun && !isTest
 
@@ -28,6 +28,7 @@ const command = defineCommand({
   },
   args: {
     ...sharedArgs,
+    ...envNameArgs,
     ...legacyRootDirArgs,
     ...getListhenArgs(),
     dotenv: {
@@ -55,6 +56,7 @@ const command = defineCommand({
     const { loadNuxtConfig } = await loadKit(cwd)
     const nuxtOptions = await loadNuxtConfig({
       cwd,
+      envName: ctx.args.envName, // c12 will fall back to NODE_ENV
       overrides: {
         dev: true,
         logLevel: ctx.args.logLevel as 'silent' | 'info' | 'verbose',
@@ -81,6 +83,7 @@ const command = defineCommand({
           logLevel: ctx.args.logLevel as 'silent' | 'info' | 'verbose',
           clear: ctx.args.clear,
           dotenv: !!ctx.args.dotenv,
+          envName: ctx.args.envName,
           loadingTemplate: nuxtOptions.devServer.loadingTemplate,
           devContext: {},
         },
@@ -167,12 +170,11 @@ async function _startSubprocess(devProxy: DevProxy, rawArgs: string[]) {
   const restart = async () => {
     // Kill previous process with restart signal
     kill('SIGHUP')
-
     // Start new process
     childProc = fork(globalThis.__nuxt_cli__!.entry!, ['_dev', ...rawArgs], {
       execArgv: [
         '--enable-source-maps',
-        process.argv.includes('--inspect') && '--inspect',
+        process.argv.find((a: string) => a.includes('--inspect')),
       ].filter(Boolean) as string[],
       env: {
         ...process.env,
@@ -244,12 +246,12 @@ function _resolveListenOptions(
     = typeof args.host === 'string'
       ? args.host
       : (args.host === true ? '' : undefined)
-      ?? process.env.NUXT_HOST
-      ?? process.env.NITRO_HOST
-      ?? process.env.HOST
+        ?? process.env.NUXT_HOST
+        ?? process.env.NITRO_HOST
+        ?? process.env.HOST
         // TODO: Default host in schema should be undefined instead of ''
-      ?? nuxtOptions._layers?.[0].config?.devServer?.host
-      ?? undefined
+        ?? nuxtOptions._layers?.[0].config?.devServer?.host
+        ?? undefined
 
   const _public: boolean | undefined
     = args.public
@@ -263,8 +265,8 @@ function _resolveListenOptions(
     || process.env.NUXT_SSL_CERT
     || process.env.NITRO_SSL_CERT
     || (typeof nuxtOptions.devServer.https !== 'boolean'
-    && nuxtOptions.devServer.https?.cert)
-    || ''
+      && nuxtOptions.devServer.https?.cert)
+      || ''
 
   const _httpsKey
     = args['https.key']
@@ -272,8 +274,8 @@ function _resolveListenOptions(
     || process.env.NUXT_SSL_KEY
     || process.env.NITRO_SSL_KEY
     || (typeof nuxtOptions.devServer.https !== 'boolean'
-    && nuxtOptions.devServer.https?.key)
-    || ''
+      && nuxtOptions.devServer.https?.key)
+      || ''
 
   const httpsEnabled
     = args.https == true
@@ -298,6 +300,6 @@ function _resolveListenOptions(
     hostname: _hostname,
     public: _public,
     https: httpsOptions,
-    baseURL: nuxtOptions.app.baseURL,
+    baseURL: nuxtOptions.app.baseURL.startsWith('./') ? nuxtOptions.app.baseURL.slice(1) : nuxtOptions.app.baseURL,
   }
 }