+++ src/commands/module/add.ts
@@ -1,7 +1,12 @@
+import * as fs from 'node:fs'
+import { homedir } from 'node:os'
+import { join } from 'node:path'
+import type { FileHandle } from 'node:fs/promises'
 import { defineCommand } from 'citty'
 import { resolve } from 'pathe'
 import consola from 'consola'
 import { addDependency } from 'nypm'
+import { joinURL } from 'ufo'
 import { $fetch } from 'ofetch'
 import { satisfies } from 'semver'
 import { updateConfig } from 'c12/update'
@@ -216,9 +221,9 @@ async function resolveModule(
 
   // Fetch package on npm
   pkgVersion = pkgVersion || 'latest'
-  const pkg = await $fetch(
-    `https://registry.npmjs.org/${pkgName}/${pkgVersion}`,
-  )
+  const pkgScope = pkgName.startsWith('@') ? pkgName.split('/')[0] : null
+  const registry = await detectNpmRegistry(pkgScope)
+  const pkg = await $fetch(joinURL(registry, `${pkgName}/${pkgVersion}`))
   const pkgDependencies = Object.assign(
     pkg.dependencies || {},
     pkg.devDependencies || {},
@@ -247,4 +252,51 @@ async function resolveModule(
     pkgName,
     pkgVersion,
   }
+}
+
+async function detectNpmRegistry(scope: string | null) {
+  if (process.env.COREPACK_NPM_REGISTRY) {
+    return process.env.COREPACK_NPM_REGISTRY
+  }
+  const userNpmrcPath = join(homedir(), '.npmrc')
+  const cwdNpmrcPath = join(process.cwd(), '.npmrc')
+  const registry = await getRegistryFromFile([cwdNpmrcPath, userNpmrcPath], scope)
+  if (registry) {
+    process.env.COREPACK_NPM_REGISTRY = registry
+  }
+  return registry || 'https://registry.npmjs.org'
+}
+
+async function getRegistryFromFile(paths: string[], scope: string | null) {
+  for (const npmrcPath of paths) {
+    let fd: FileHandle | undefined
+    try {
+      fd = await fs.promises.open(npmrcPath, 'r')
+      if (await fd.stat().then(r => r.isFile())) {
+        const npmrcContent = await fd.readFile('utf-8')
+
+        if (scope) {
+          const scopedRegex = new RegExp(`^${scope}:registry=(.+)$`, 'm')
+          const scopedMatch = npmrcContent.match(scopedRegex)
+          if (scopedMatch) {
+            return scopedMatch[1].trim()
+          }
+        }
+
+        // If no scoped registry found or no scope provided, look for the default registry
+        const defaultRegex = /^\s*registry=(.+)$/m
+        const defaultMatch = npmrcContent.match(defaultRegex)
+        if (defaultMatch) {
+          return defaultMatch[1].trim()
+        }
+      }
+    }
+    catch {
+    // swallow errors as file does not exist
+    }
+    finally {
+      await fd?.close()
+    }
+  }
+  return null
 }